# services/go_nn/src/api/irest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Нейронные сети по детекции объектов на изображениях",
        "title": "Нейросети API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/tasks/one": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Task with one image",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Task",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File image",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task one image",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}